test_hypergraph_agent.py
import unittest
from hypergraph_agent import HypergraphAgent

class TestHypergraphAgent(unittest.TestCase):
    def setUp(self):
        self.agent = HypergraphAgent()

    def test_agent_initialization(self):
        """
        Testa a inicialização do HypergraphAgent.
        """
        self.assertIsInstance(self.agent, HypergraphAgent)
        # Exemplo: self.assertEqual(self.agent.some_attribute, expected_value)

    def test_some_functionality(self):
        """
        Testa alguma funcionalidade do HypergraphAgent.
        """
        # Adicione asserções para testar funcionalidades específicas
        result = self.agent.some_method()
        expected_result = "expected_value"  # Define o resultado esperado
        self.assertEqual(result, expected_result)

    # Adicione mais métodos de teste conforme necessário

if __name__ == '__main__':
    unittest.main()